implement: add create meme api

See merge request flexstack.ai/membots-blockchain-be!8

+++ internal/dto/nft.go
@@ -14,9 +14,9 @@ type Nft struct {
 type Tier struct {
 	MemeId          string   `json:"memeId"`
 	NftId           string   `json:"nftId"`
-	Rank            uint     `json:"rank"`
+	Rank            uint64   `json:"rank"`
 	AmountThreshold string   `json:"amountThreshold"`
-	TokenIdLower    uint     `json:"tokenIdLower"`
+	TokenIdLower    uint64   `json:"tokenIdLower"`
 	TokenIdUpper    string   `json:"tokenIdUpper"`
 	TokenIdURLs     []string `json:"tokenIdURLs"`
 }
